+++ core/class/rosee.class.php
@@ -223,21 +223,21 @@ class rosee extends eqLogic {
             $rosee_point = round(($rosee), 1);
         
             // Calcul visibilité Alerte Point de rosée
-                    if ($rosee_point >= 0.0) {
-                        $visible_Rosee = 1;
-                    } else {
-                        $visible_Rosee = 0;
-                    }
+                  //  if ($rosee_point >= 0.0) {
+                 //       $visible_Rosee = 1;
+                //    } else {
+                //        $visible_Rosee = 0;
+                //    }
 
             log::add('rosee', 'debug', '========= CALCUL DU POINT DE ROSEE ========');
             log::add('rosee', 'debug', 'Point de Rosée : ' . $rosee_point);
-            log::add('rosee', 'debug', 'Visibilité Point de Rosée : ' . $visible_Rosee);
+        //    log::add('rosee', 'debug', 'Visibilité Point de Rosée : ' . $visible_Rosee);
         
         // Calcul de l'alerte rosée en fonction du seuil d'alerte
             $frost_alert_rosee = $temperature - $rosee_point;
                 log::add('rosee', 'debug', 'Calcul point de rosee (Température - point de Rosée) : ' . $frost_alert_rosee );
         
-            if ($visible_Rosee == 1) {
+           // if ($visible_Rosee == 1) {
                 if (($frost_alert_rosee) <= $dpr) {
                     $alert_r = 1;
                     log::add('rosee', 'debug', 'RESULTAT Calcul point de rosee (Calcul point de Rosée  <= Seuil DPR)');
@@ -245,10 +245,10 @@ class rosee extends eqLogic {
                     $alert_r = 0;
                     log::add('rosee', 'debug', 'RESULTAT Calcul point de rosee (Calcul point de Rosée  > Seuil DPR)');
                 }
-            } else {
-                $alert_r = 0;
-                log::add('rosee', 'debug', 'AUCUN Calcul point de Rosée car Visibilité Point de Rosée = 0');
-            }
+          //  } else {
+              //  $alert_r = 0;
+             //   log::add('rosee', 'debug', 'AUCUN Calcul point de Rosée car Visibilité Point de Rosée = 0');
+          //  }
 		  log::add('rosee', 'debug', 'Etat alerte rosée ' . $alert_r);
         
         
@@ -265,21 +265,21 @@ class rosee extends eqLogic {
             $frost_point = round(($frost), 1);
         
             // Calcul visibilité Alerte Point de Givrage
-                    if($frost_point < 0.0) {
-                        $visible_Frost = 1;
-                    } else {
-                        $visible_Frost = 0;
-                    }
+                   // if($frost_point < 0.0) {
+                     //   $visible_Frost = 1;
+                    //} else {
+                    //    $visible_Frost = 0;
+                //    }
 
             log::add('rosee', 'debug', '======== CALCUL DU POINT DE GIVRAGE =======');
             log::add('rosee', 'debug', 'Point de Givrage :' . $frost_point.' °C');
-            log::add('rosee', 'debug', 'Visibilité Point de Givrage : ' . $visible_Frost);
+          //  log::add('rosee', 'debug', 'Visibilité Point de Givrage : ' . $visible_Frost);
 
         // Calcul de l'alerte givrage en fonction du seuil d'alerte
             $frost_alert_givrage = $temperature - $frost_point;
                 log::add('rosee', 'debug', 'Calcul point de givrage (Température - point de givrage) : ' . $frost_alert_givrage);
         
-            if ($visible_Frost == 1) {
+           // if ($visible_Frost == 1) {
                 if (($frost_alert_givrage) <= $dpr) {
                     $alert_g = 1;
                     log::add('rosee', 'debug', 'RESULTAT Calcul point de givrage (Calcul point de givrage  <= Seuil DPR)');
@@ -287,10 +287,10 @@ class rosee extends eqLogic {
                     $alert_g = 0;
                     log::add('rosee', 'debug', 'RESULTAT Calcul point de givrage (Calcul point de givrage  > Seuil DPR)');
                 }
-            } else {
-                $alert_g = 0;
-                log::add('rosee', 'debug', 'AUCUN Calcul point de givrage car Visibilité Point de Givrage = 0');
-            }
+            //} else {
+              //  $alert_g = 0;
+            //    log::add('rosee', 'debug', 'AUCUN Calcul point de givrage car Visibilité Point de Givrage = 0');
+            //}
         
             log::add('rosee', 'debug', 'Etat alerte gel : ' . $alert_g);