+++ docs/fr_FR/changelog.md
@@ -9,6 +9,7 @@ description: Ce plugin permet d'obtenir le point de rosée pour savoir si l'herb
 # Version 3.3.2
 
 - Ajout de log supplémentaire en mode DEBUG
+- Suppression visibilité pour le calcul du point de rosée et du point de givrage
 
 # Version 3.3.1